slight refactor of lscale calculation.
Main changes are:

- The subroutine calculate_lscale now operates on plain arrays
  instead of EAM's derived types (state and pbuf).

- The diagnosed Lscales are taken out of the state variable
  and declared as local array in subroutine tphysbc for now.

These changes are motivated by the reasoning that in an AMR setting,
we will be calculating Lscales on the refined/adjusted grid;
at this point, it's not yet clear what kind data structures
will be used, hence for now, we make the lscale_mod.F90 more
plain and general.

1-hr test simulations were performed using the "mac_simprad"
configuration for physics and the default IOP file with omega,
using

 1) SHOC as the turb parameterization; debug build;
 2) CLUBB as the turb parameterization; default optimization;

In both groups, simulations were performed using dz = ~100m or 10m,
dt = 60s. I confirmed that before and after the refactor,
both the time integration (i.e., the "nstep, te" lines in atm.log)
and the history output were BFB identical.